+++ lemon/dashboard/settings.py
@@ -1,7 +1,0 @@
-from django.conf import settings
-
-
-CONFIG = {
-    'STATE': ((), ()),
-}
-CONFIG.update(getattr(settings, 'DASHBOARD_CONFIG', {}))