+++ lemon/extradmin/__init__.py
@@ -1,5 +1,0 @@
-from django.contrib.admin import autodiscover, HORIZONTAL, VERTICAL
-
-from lemon.extradmin.options import ModelAdmin
-from lemon.extradmin.options import StackedInline, TabularInline
-from lemon.extradmin.sites import AdminSite, site