+++ lemon/extradmin/settings.py
@@ -1,7 +1,0 @@
-from django.conf import settings
-
-
-CONFIG = {
-    'MARKUP_WIDGET': None,
-}
-CONFIG.update(getattr(settings, 'EXTRADMIN_CONFIG', {}))