+++ lemon/metatags/__init__.py
@@ -1,1 +1,0 @@
-from lemon.metatags.sites import MetatagsSite, site