+++ lemon/robots/tests/urls.py
@@ -1,6 +1,0 @@
-from django.conf.urls import patterns, include, url
-
-
-urlpatterns = patterns('',
-    url(r'^robots\.txt', include('lemon.robots.urls')),
-)