+++ lemon/sitemaps/__init__.py
@@ -1,1 +1,0 @@
-from lemon.sitemaps.sites import SitemapSite, site