+++ lemon/sitemaps/tests/urls.py
@@ -1,6 +1,0 @@
-from django.conf.urls import patterns, include, url
-
-
-urlpatterns = patterns('',
-    url(r'^sitemap\.xml', include('lemon.sitemaps.urls')),
-)