+++ HISTORY.rst
@@ -1,6 +1,13 @@
 History
 =======
 
+0.7
+---
+
+* Rename lemon.extradmin app to lemon
+* Move all apps except extradmin to their own separate distributions
+* Add ability to register custom admin sections per app using AppAdmin
+
 0.6.2
 -----